feat: v0.4.0 — OpenXR bring-up, EFX reverb, gamepad bindings, headlamp investigation, static analysis fixes
OpenXR / VR plumbing (XR.cpp / XR.h):
- Dynamic-load OpenXR via LoadLibrary; game runs without VR runtime installed
- Steps 1–3 complete: loader probe, xrCreateInstance/xrGetSystem, GL-bound
  xrCreateSession + xrPollEvent session state pump
- Q/E snap-turn (±30°) wired in flatscreen as VR-comfort primitive

Audio (Audio_OpenAL.cpp):
- EFX reverb zones: per-area presets via alGenEffects (forest/outdoor defaults)
- HRTF toggle via ALC_HRTF_SOFT; occlusion raycast stub wired
- Ambient fade moved to dedicated update path; menu ambient snap-to fixed

Renderer / shaders (RendererGL.cpp, renderd3d.cpp, basic.frag, basic.vert):
- Perspective-correct UV via noperspective vTexCoordR/vRhw; fixes terrain
  swimming on NVIDIA when camera rotates
- Depth-based LOD for foliage replaces screen-space derivatives; eliminates
  headbob-induced strobe/shimmer
- Alpha test at LOD=0 for precise binary keying; colour sampling at full LOD
- Debug visualization modes 0–13 (F8) for renderer diagnostics
- Hue-preserving brightness clamp (scale-down vs hard clip) fixes warm-colour
  saturation shift at high uBrightness

Gameplay / input (Hunt2.cpp, Game.cpp, Gamepad.cpp):
- Configurable gamepad bindings persisted to controls.cfg
- VFS::Init called from SDL main(); mod pack mounting at launch
- Hysteresis footstep gate (was firing ~4× per cycle at high framerate)
- Multi-jump prevention: kfJump gated on rising edge only

64-bit / static analysis fixes (Loading.cpp, Interface.cpp, mathematics.cpp, Resources.cpp):
- WndProc signature: LONG/UINT → LRESULT/WPARAM/LPARAM
- GDI handle casts: SelectObject/GetStockObject return HGDIOBJ; added
  explicit (HBITMAP)/(HFONT)/(HBRUSH) casts throughout
- strlen results cast to (int) before Win32 text functions
- HeapSize (SIZE_T) cast to (DWORD) before accumulation into DWORD counter
- Float math: fabs→fabsf, sqrt→sqrtf, sin/cos→sinf/cosf, double literals
  given f suffix, explicit (float) casts on intentional double→float narrowing

Project files: VS solution/vcxproj added; code_analysis.ruleset committed

+++ Interface.cpp
@@ -50,7 +50,7 @@ void wait_mouse_release()
 int GetTextW(HDC hdc, LPSTR s)
 {
   SIZE sz;
-  GetTextExtentPoint(hdc, s, strlen(s), &sz);
+  GetTextExtentPoint(hdc, s, (int)strlen(s), &sz);
   return sz.cx;
 }
 
@@ -59,10 +59,10 @@ void PrintText(LPSTR s, int x, int y, int rgb)
   HBITMAP hbmpOld = (HBITMAP)SelectObject(hdcCMain,hbmpVideoBuf);   
   SetBkMode(hdcCMain, TRANSPARENT);     
    
-  SetTextColor(hdcCMain, 0x00000000);  
-  TextOut(hdcCMain, x+1, y+1, s, strlen(s));
+  SetTextColor(hdcCMain, 0x00000000);
+  TextOut(hdcCMain, x+1, y+1, s, (int)strlen(s));
   SetTextColor(hdcCMain, rgb);
-  TextOut(hdcCMain, x, y, s, strlen(s));
+  TextOut(hdcCMain, x, y, s, (int)strlen(s));
 
   SelectObject(hdcCMain,hbmpOld);		  
 }
